Lucene.Net 2.0 "Alpha"
git-svn-id: https://svn.apache.org/repos/asf/incubator/lucene.net/trunk@432239 13f79535-47bb-0310-9956-ffa450edef68

+++ C#/src/Lucene.Net/Store/RAMDirectory.cs
@@ -101,11 +101,11 @@ namespace Lucene.Net.Store
 				// read current file
 				IndexInput is_Renamed = dir.OpenInput(files[i]);
 				// and copy to ram disk
-				int len = (int) is_Renamed.Length();
-				int readCount = 0;
+				long len = (int) is_Renamed.Length();
+				long readCount = 0;
 				while (readCount < len)
 				{
-					int toRead = readCount + BufferedIndexOutput.BUFFER_SIZE > len?len - readCount : BufferedIndexOutput.BUFFER_SIZE;
+					int toRead = readCount + BufferedIndexOutput.BUFFER_SIZE > len ? (int) (len - readCount) : BufferedIndexOutput.BUFFER_SIZE;
 					is_Renamed.ReadBytes(buf, 0, toRead);
 					os.WriteBytes(buf, toRead);
 					readCount += toRead;
